requirements.txt
Django>=3.2
psycopg2-binary
django-redis
djangorestframework
django-allauth
celery[redis]
gunicorn
